+++ src/Websites/Websites.Test/ScenarioTests/WebAppTests.ps1
@@ -1464,6 +1464,7 @@ function Test-WebAppPublishingProfile
 	}
 }
 
+# Keep existing test to ensure backwards compatibility with existing behavior
 function Test-PublishAzureWebAppFromZip
 {
 	# Setup
@@ -1494,6 +1495,7 @@ function Test-PublishAzureWebAppFromZip
 	}
 }
 
+# Keep existing test to ensure backwards compatibility with existing behavior
 function Test-PublishAzureWebAppFromWar
 {
 	# Setup
@@ -1523,6 +1525,46 @@ function Test-PublishAzureWebAppFromWar
 
 		$warPath = Join-Path $ResourcesPath "HelloJava.war"
 		$publishedApp = Publish-AzWebApp -ResourceGroupName $rgname -Name $appName -ArchivePath $warPath -Force
+
+		Assert-NotNull $publishedApp
+	}
+	finally
+	{
+		# Cleanup
+		Remove-AzureRmResourceGroup -Name $rgname -Force
+	}
+}
+
+# New tests for PublishAzureWebApp to test deploying against newer Onedeploy endpoint
+function Test-PublishAzureWebAppOnedeploy
+{
+	# Setup
+	$rgname = Get-ResourceGroupName
+	$appName = Get-WebsiteName
+	$location = Get-WebLocation
+	$planName = Get-WebHostPlanName
+	$tier = "Shared"
+
+	try
+	{
+		#Setup
+		New-AzureRmResourceGroup -Name $rgname -Location $location
+		$serverFarm = New-AzureRmAppServicePlan -ResourceGroupName $rgname -Name  $planName -Location  $location -Tier $tier
+		
+		# Create new web app
+		$webapp = New-AzureRmWebApp -ResourceGroupName $rgname -Name $appName -Location $location -AppServicePlan $planName 
+		
+		#Configuring jdk and web container
+        # Set Java runtime to 1.8 | Tomcat. In order to deploy war, site should be configured to run with stack = TOMCAT 
+		# or JBOSSEAP (only availble on Linux). In this test case, it creates Windows app. 
+		$javaVersion="1.8"
+        $javaContainer="TOMCAT"
+        $javaContainerVersion="8.5"
+        $PropertiesObject = @{javaVersion = $javaVersion;javaContainer = $javaContainer;javaContainerVersion = $javaContainerVersion}
+        New-AzResource -PropertyObject $PropertiesObject -ResourceGroupName $rgname -ResourceType Microsoft.Web/sites/config -ResourceName "$appName/web" -ApiVersion 2018-02-01 -Force
+
+		$warPath = Join-Path $ResourcesPath "HelloJava.war"
+		$publishedApp = Publish-AzWebApp -ResourceGroupName $rgname -Name $appName -ArchivePath $warPath -Type war -Clean $true -TargetPath /home/site/wwwroot/webapps/ROOT -Force
 
 		Assert-NotNull $publishedApp
 	}