+++ src/Websites/Websites/ChangeLog.md
@@ -21,6 +21,7 @@
 
 ## Version 2.15.1
 * Used AAD Auth instead of Basic Auth for PublishAzureWebApps
+* Add support for OneDeploy API in PublishAzureWebApps while maintaining backwards compatibility with existing behavior
 
 ## Version 2.15.0
 * Fixed Tag parameter issues with ASE for `New-AzWebApp`